refactor: separate social links into its own table
